refactor: navigators 폴더 생성
