+++ src/App.tsx
@@ -1,23 +1,10 @@
-import React from 'react';
-import {createStackNavigator} from '@react-navigation/stack';
 import {SafeAreaView, StyleSheet} from 'react-native';
-import {NavigationContainer} from '@react-navigation/native';
-import Home from './screens/Home';
-
-const Stack = createStackNavigator();
+import Navigation from './navigators/Navigation';
 
 const App = () => {
   return (
     <SafeAreaView style={styles.container}>
-      <NavigationContainer>
-        <Stack.Navigator
-          initialRouteName="Home"
-          screenOptions={{
-            headerShown: false,
-          }}>
-          <Stack.Screen name="Home" component={Home} />
-        </Stack.Navigator>
-      </NavigationContainer>
+      <Navigation />
     </SafeAreaView>
   );
 };
