+++ src/navigators/Navigation.tsx
@@ -1,0 +1,12 @@
+import {NavigationContainer} from '@react-navigation/native';
+import StackNavigator from './StackNavigator';
+
+const Navigation = () => {
+  return (
+    <NavigationContainer>
+      <StackNavigator />
+    </NavigationContainer>
+  );
+};
+
+export default Navigation;
