+++ src/navigators/StackNavigator.tsx
@@ -1,0 +1,18 @@
+import {createStackNavigator} from '@react-navigation/stack';
+import Home from '../screens/Home';
+
+const Stack = createStackNavigator();
+
+const StackNavigator = () => {
+  return (
+    <Stack.Navigator
+      initialRouteName="Home"
+      screenOptions={{
+        headerShown: false,
+      }}>
+      <Stack.Screen name="Home" component={Home} />
+    </Stack.Navigator>
+  );
+};
+
+export default StackNavigator;
