+++ src/screens/Home.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import {StyleSheet, View} from 'react-native';
 import WebView from 'react-native-webview';
